// src/app/model/request.model.ts
//#region CLASS

export class RequestModel {
	Data?: string;

	ID?: number;
	UserID?: number;
	ClientDeviceID?: string;
	IPAddress?: string;
	Culture?: string;
	ServerApplicationVersion?: string;
	ClientApplicationVersion?: string;
	ClientOperatingSystem?: string;
	ClientDeviceModel?: string;
	ClientDeviceBrand?: string;
	ClientDeviceSerie?: string;
	ClientDeviceBuildID?: string;
	ClientDeviceProcessor?: string;
	ClientDeviceStable?: string;
	ClientDeviceBrowser?: string;
	ClientDeviceEngine?: string;
	ClientDeviceTool?: string;

	search?: string;
}

//#endregion